-fix crash in VS2005 - rand_source_init is overwritting past end of buffer
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@7739 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/contrib/srtp/crypto/rng/rand_source.c
@@ -109,7 +109,7 @@ rand_source_get_octet_string(void *dest, uint32_t len) {
   if (read(dev_random_fdes, dest, len) != len)
     return err_status_fail;
 #elif (_MSC_VER >= 1400)
-  unsigned int *dst = dest;
+  uint8_t *dst = dest;
   while (len)
   {
       unsigned int val = 0;
